restructured stuff and removed som files
