+++ Makefile
@@ -2,7 +2,7 @@ init:
 	python -c"from minitwit import init_db; init_db()"
 
 build:
-	gcc flag_tool.c -l sqlite3 -o flag_tool
+	gcc tools/flag_tool.c -l sqlite3 -o bin/flag_tool
 
 clean:
 	rm flag_tool
